FATFileSystem -Fixed API integer widths

+++ libraries/fs/fat/FATFileSystem.h
@@ -84,10 +84,10 @@ public:
 
     virtual int disk_initialize() { return 0; }
     virtual int disk_status() { return 0; }
-    virtual int disk_read(uint8_t * buffer, uint64_t sector, uint8_t count) = 0;
-    virtual int disk_write(const uint8_t * buffer, uint64_t sector, uint8_t count) = 0;
+    virtual int disk_read(uint8_t *buffer, uint32_t sector, uint32_t count) = 0;
+    virtual int disk_write(const uint8_t *buffer, uint32_t sector, uint32_t count) = 0;
     virtual int disk_sync() { return 0; }
-    virtual uint64_t disk_sectors() = 0;
+    virtual uint32_t disk_sectors() = 0;
 
 };
